Rename package, init mkdocs, pytest, pyproject.toml, and delete obsolete files
Removed an outdated cached file that is no longer in use. This cleanup ensures the repository remains free of unnecessary clutter.
