--- docs/source/api_reference.rst
@@ -1,36 +1,0 @@
-API Reference
-=============
-
-landlens_db.geoclasses
-----------------------
-
-.. automodule:: landlens_db.geoclasses.geoimageframe
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-landlens_db.handlers
---------------------
-
-.. automodule:: landlens_db.handlers.cloud
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-.. automodule:: landlens_db.handlers.image
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-.. automodule:: landlens_db.handlers.db
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-landlens_db.snap
---------------------
-
-.. automodule:: landlens_db.process.snap
-   :members:
-   :undoc-members:
-   :show-inheritance:
