--- docs/source/index.rst
@@ -1,26 +1,0 @@
-.. landlens_db documentation master file, created by
-   sphinx-quickstart on Wed Mar  1 21:12:52 2023.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-landlens_db
-===========
-
-Welcome to landlens_db's documentation!
-
-Contents:
----------
-
-.. toctree::
-   :maxdepth: 3
-
-   introduction
-   getting_started
-   api_reference
-
-Indices and Tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
