+++ pytest.ini
@@ -1,0 +1,6 @@
+[pytest]
+minversion = 7.0
+addopts = -ra -q
+testpaths =
+    tests
+python_files = test_*.py *_test.py
