--- requirements-dev.txt
@@ -1,12 +1,0 @@
-# Include main requirements
--r requirements.txt
-
-# Development requirements
-pytest==8.0.0
-jupyter~=1.0.0
-pre-commit~=3.6.0
-python-dotenv==1.0.1
-sphinx==7.2.6
-sphinx_rtd_theme==2.0.0
-setuptools==69.0.3
-wheel~=0.42.0
