--- requirements.txt
@@ -1,21 +1,0 @@
-numpy==1.26.4
-geopandas==0.14.3
-geoalchemy2==0.14.7
-setuptools==69.2.0
-# Dependencies automatically installed
-SQLAlchemy==2.0.38
-pandas==2.2.3
-shapely==2.0.7
-fiona==1.10.1
-pyproj==3.7.1
-packaging==24.2
-python-dateutil==2.9.0.post0
-pytz==2025.1
-tzdata==2025.1
-attrs==25.1.0
-certifi==2025.1.31
-click==8.1.8
-click-plugins==1.1.1
-cligj==0.7.2
-six==1.17.0
-typing-extensions==4.12.2
